Moved Parser utils into their own directory, added AuthorParser. Pending implementation. Removed debug output.
