+++ src/app/parser/impl/DagbladetParser.ts
@@ -3,16 +3,11 @@ import Parser from 'base/Parser';
 
 import GenericParser from './GenericParser';
 
-import { getOpenGraphModifiedTime } from '../VersionParser';
-
-import * as app from 'app/util/applib';
-
-const log = app.createLog();
+import { getOpenGraphModifiedTime } from '../util/VersionParser';
 
 export default class DagbladetParser extends GenericParser {
 
 	protected parseVersion() : string {
-		log('parseVersion');
 		return getOpenGraphModifiedTime(this.select);
 	}
