+++ src/app/parser/impl/GenericParser.ts
@@ -8,16 +8,11 @@ import Parser from 'base/Parser';
 
 import BaseParser from '../BaseParser';
 
-import * as app from 'app/util/applib';
-
-const log = app.createLog();
-
 export default class GenericParser extends BaseParser implements Parser {
 
 	protected select : Cheerio;
 
 	parse(rawHTML : string, url : ArticleURL) : Promise <Article> {
-		log('parse()');
 		this.select = Cheerio.load(rawHTML);
 		return super.parse(rawHTML, url);
 	}
