+++ src/app/parser/util/AuthorParser.ts
@@ -1,0 +1,12 @@
+import * as Cheerio from 'cheerio';
+
+import ArticleAuthor from 'base/ArticleAuthor';
+
+/**
+ * Facebook's OpenGraph scheme
+ * looks for <meta property="article:modified_time" content="...">
+ * and <meta property="article:bylineEmail" content="...">
+ */
+export function getOpenGraphAuthors(select : Cheerio) : ArticleAuthor[] {
+	return [];
+}
